+++ src/realm/util/file_mapper.cpp
@@ -928,9 +928,13 @@ void msync(FileDesc fd, void* addr, size_t size)
     return;
 #else
     static_cast<void>(fd);
-    if (::msync(addr, size, MS_SYNC) != 0) {
+    int retries_left = 1000;
+    while (::msync(addr, size, MS_SYNC) != 0) {
         int err = errno; // Eliminate any risk of clobbering
-        throw std::system_error(err, std::system_category(), "msync() failed");
+        if (--retries_left < 0)
+            throw std::system_error(err, std::system_category(), "msync() retries exhausted");
+        if (err != EINTR)
+            throw std::system_error(err, std::system_category(), "msync() failed");
     }
 #endif
 }